+++ packages/lib/src/extras/SplashScreen.ts
@@ -9,8 +9,8 @@ import { wait } from "inks2d/utils";
 export class SplashScreen extends Scene {
 	private _loader: Rectangle = new Rectangle(1080, 1920, "black");
 	private _assetsToLoad: string[];
-	private _image: DisplayObject = new Rectangle(0, 0);
-	private _imagePath: string;
+	private _logo: DisplayObject = new Rectangle(0, 0);
+	private _splashImage: string | DisplayObject;
 	private _callback: () => void;
 	private _forceClick: boolean;
 	// @ts-ignore
@@ -22,14 +22,14 @@ export class SplashScreen extends Scene {
 		assetsToLoad: string[],
 		callback: () => void,
 		yOffset: number = 0,
-		imagePath: string,
+		splashImage: string | DisplayObject,
 		forceClick: boolean = false,
 		ctaText: string = "TAP TO START",
 	) {
 		super();
 
 		this._assetsToLoad = assetsToLoad;
-		this._imagePath = imagePath;
+		this._splashImage = splashImage;
 		this._callback = callback;
 		this._forceClick = forceClick;
 		this._ctaText = ctaText;
@@ -45,37 +45,45 @@ export class SplashScreen extends Scene {
 		this._loader.height = this._g.stage.height;
 		this._g.stage.addChild(this._loader);
 
-		if (!this._imagePath) {
+		if (!this._splashImage) {
 			this._assetsToLoad && this.loadAssets();
 			return;
 		}
 
-		this._g.loader.onComplete = () => {
-			const maxWidth = this._g.stage.width / 2.5;
-			const maxHeight = this._g.stage.height / 2.5;
-			const minSize = Math.min(Math.min(maxWidth, maxHeight), 300);
-			const targetArea = minSize * minSize;
+		if (typeof this._splashImage === "string") {
+			this._g.loader.onComplete = () => {
+				const maxWidth = this._g.stage.width / 2.5;
+				const maxHeight = this._g.stage.height / 2.5;
+				const minSize = Math.min(Math.min(maxWidth, maxHeight), 300);
+				const targetArea = minSize * minSize;
 
-			this._image = new Sprite(this._g.loader.store[this._imagePath]);
-			const newWidth = Math.sqrt(
-				(this._image.width / this._image.height) * targetArea,
-			);
-			const newHeight = targetArea / newWidth;
-			this._image.width = Math.round(newWidth);
-			this._image.height = Math.round(
-				newHeight - (this._image.width - newWidth),
-			);
-			this._assetsToLoad && this.loadAssets();
-		};
-		this._g.loader.load([this._imagePath]);
+				this._logo = new Sprite(
+					this._g.loader.store[this._splashImage as string],
+				);
+				const newWidth = Math.sqrt(
+					(this._logo.width / this._logo.height) * targetArea,
+				);
+				const newHeight = targetArea / newWidth;
+				this._logo.width = Math.round(newWidth);
+				this._logo.height = Math.round(
+					newHeight - (this._logo.width - newWidth),
+				);
+				this._assetsToLoad && this.loadAssets();
+			};
+			this._g.loader.load([this._splashImage]);
+			return;
+		}
+
+		this._logo = this._splashImage;
+		this._assetsToLoad && this.loadAssets();
 	}
 
 	private loadAssets(): void {
 		const maxWidth = this._g.stage.width / 2.5;
 
-		this._image.pivot.x = this._image.pivot.y = 0.5;
-		this._loader.putCenter(this._image);
-		this._loader.addChild(this._image);
+		this._logo.pivot.x = this._logo.pivot.y = 0.5;
+		this._loader.putCenter(this._logo);
+		this._loader.addChild(this._logo);
 
 		const back = new Rectangle(maxWidth, 25, "gray", "black", 2, 5);
 		back.pivot.x = 0;
@@ -84,7 +92,7 @@ export class SplashScreen extends Scene {
 		back.position.x = 540 - 210;
 		back.position.y = 960;
 		this.updateObjPos(back);
-		back.position.y += this._image.height / 2 + back.height + this._yOffset;
+		back.position.y += this._logo.height / 2 + back.height + this._yOffset;
 		this._loader.addChild(back);
 
 		const front = new Rectangle(0, 25, "#ddd", "none", 2, 1);
@@ -93,7 +101,7 @@ export class SplashScreen extends Scene {
 		front.position.x = 540 - 210;
 		front.position.y = 960;
 		this.updateObjPos(front);
-		front.position.y += this._image.height / 2 + back.height + this._yOffset;
+		front.position.y += this._logo.height / 2 + back.height + this._yOffset;
 		this._loader.addChild(front);
 
 		this._g.loader.onUpdate = (loaded, total) => {
@@ -134,7 +142,7 @@ export class SplashScreen extends Scene {
 		start.size = (this._g.stage.width * start.size) / 1080;
 		this.updateObjPos(start);
 		start.position.y +=
-			this._image.height / 2 +
+			this._logo.height / 2 +
 			this._loader.children[1].height * 3.5 +
 			this._yOffset;
 		this._loader.addChild(start);